requirements.txt
streamlit
cohere
python-dotenv
pdfplumber
pandas
